# services/core/docs/swagger.yaml
basePath: /api/v1
host: localhost:8080
info:
  contact: {}
  description: B 站风格全栈视频平台核心 API（稿件/评论/收藏/用户）
  title: mybilibili Core API
  version: "1.0"
paths:
...
  /favorites/check:
    get:
      description: 按稿件 id 或视频 id 判断当前用户是否已收藏；兼容 manuscript_id / manuscriptId / vid
        三种参数名
      parameters:
      - description: 稿件 id
        in: query
        name: manuscript_id
        type: integer
      - description: 稿件 id（驼峰写法）
        in: query
        name: manuscriptId
        type: integer
      - description: 视频 id
        in: query
        name: vid
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Favorite check result
          schema:
            type: string
        "401":
          description: 未登录
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: 检查是否已收藏
      tags:
      - favorite
  /favorites/list:
    get:
      description: 返回当前登录用户收藏过的全部稿件，按收藏时间倒序
      parameters:
      - description: 页码（默认 1）
        in: query
        name: page
        type: integer
      - description: 每页条数（默认 30）
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Favorite list
          schema:
            type: string
        "401":
          description: 未登录
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: 我的收藏列表（平铺）
      tags:
      - favorite
  /favorites/manuscript/{id}:
    get:
      description: 按稿件 id 取该稿件被加入的全部收藏夹
      parameters:
      - description: 稿件 id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Manuscript folders
          schema:
            type: string
        "400":
          description: invalid manuscript id
          schema:
            type: string
      summary: 稿件所属收藏夹列表
      tags:
      - favorite
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"